ajustando o código para verificar a existência da base de dados

+++ experiments/exp2.sh
@@ -33,7 +33,8 @@ echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 1
 
 echo "Start cloud task manager"
-rm cloud_task_manager/db/tasks.db
+[ -d "cloud_task_manager/db/tasks.db" ]  && rm cloud_task_manager/db/tasks.db
+
 python -u -m cloud_task_manager.service_cloud_ml > logs_${LOG_TIMESTAMP}/service_cloud_ml.log 2>&1 &
 echo $! >>  logs_${LOG_TIMESTAMP}/pids
 sleep 1
